linuxOperation/requirement.txt
#asn1crypto==0.24.0
amqp==2.3.2
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.1.1
kombu==4.2.1
cffi==1.11.2
chardet==3.0.4
Django==1.11.3
django-logtail==0.3.0
django-redis==4.8.0
enum34==1.1.6
freeze==1.0.10
gevent==1.2.2
greenlet==0.4.14
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
IPy==0.83
M2Crypto==0.26.2
MySQL-python==1.2.5
passlib==1.7.1
phpserialize==1.3
psutil==5.4.0
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
PyNaCl==1.2.1
pytz==2017.2
redis==2.10.6
six==1.11.0
typing==3.6.2
jsonfield==2.0.2
python-dateutil==2.7.3
vine==1.1.4
xlrd==1.0.0
xlwt==1.1.2
celery==4.2.1
netifaces==0.10.7
pypinyin==0.33.0
